Create SQL Queries
Creating a SQL table in the database to hold the data from the csv file and then we run a query to load the data into the newly created table. The following queries are pulling varying amounts of information that can then be sent to python for a more thorough analysis.